Refactor code into command-specific Go files
This moves all command processing code out of cmd/fdo_client/client.go
and moves it into per-command files under the cmd package. Removes the
cmd/fdo_client subdirectory and updates the command examples in the
README file.
