+++ Makefile
@@ -1,11 +1,10 @@
 # Variables
 BINARY_NAME = fdo_client
-BUILD_DIR = ./cmd/fdo_client
 CRED_NAME = cred.bin
 
 # Build the Go project
 build:
-	go build -o $(BINARY_NAME) $(BUILD_DIR)
+	go build -o $(BINARY_NAME)
 
 # Clean up the binary
 clean:
