+++ README.md
@@ -24,45 +24,72 @@ To update your dependencies, simply run the script:
 The client application can be built with `make build` or `go build` directly,
 
 ```console
-$ make build or go build -o fdo_client ./cmd/fdo_client
+$ make build or go build -o fdo_client
 $ ./fdo_client
 
-Usage:
-  fdo_client [--] [options]
+FIDO Device Onboard Client
 
-Client options:
-  -blob string
-        File path of device credential blob (default "cred.bin")
-  -cipher suite
-        Name of cipher suite to use for encryption (see usage) (default "A128GCM")
-  -debug
-        Print HTTP contents
-  -di URL
-        HTTP base URL for DI server
-  -di-key string
-        Key for device credential [options: ec256, ec384, rsa2048, rsa3072] (default "ec384")
-  -di-key-enc string
-        Public key encoding to use for manufacturer key [x509,x5chain,cose] (default "x509")
-  -download dir
-        A dir to download files into (FSIM disabled if empty)
-  -echo-commands
-        Echo all commands received to stdout (FSIM disabled if false)
-  -insecure-tls
-        Skip TLS certificate verification
-  -kex suite
-        Name of cipher suite to use for key exchange (see usage) (default "ECDH384")
-  -print
-        Print device credential blob and stop
-  -rv-only
-        Perform TO1 then stop
-  -resale
-        Perform resale
-  -tpm path
-        Use a TPM at path for device credential secrets
-  -upload files
-        List of dirs and files to upload files from, comma-separated and/or flag provided multiple times (FSIM disabled if empty)
-  -wget-dir dir
-        A dir to wget files into (FSIM disabled if empty)
+Usage:
+  fdo_client [command]
+
+Available Commands:
+  device-init Run device initialization (DI)
+  help        Help about any command
+  onboard     Run FDO TO1 and TO2 onboarding
+  print       Print device credential blob and exit
+
+Flags:
+      --blob string   File path of device credential blob
+      --debug         Print HTTP contents
+  -h, --help          help for fdo_client
+      --tpm string    Use a TPM at path for device credential secrets
+
+Use "fdo_client [command] --help" for more information about a command.
+
+$ ./fdo_client device-init -h
+
+Run device initialization (DI)
+
+Usage:
+  fdo_client device-init [flags]
+
+Flags:
+      --di string                   HTTP base URL for DI server (default "http://127.0.0.1:8080")
+      --di-device-info string       Device information for device credentials, if not specified, it'll be gathered from the system
+      --di-device-info-mac string   Mac-address's iface e.g. eth0 for device credentials
+      --di-key string               Key for device credential [options: ec256, ec384, rsa2048, rsa3072]
+      --di-key-enc string           Public key encoding to use for manufacturer key [x509,x5chain,cose] (default "x509")
+  -h, --help                        help for device-init
+      --insecure-tls                Skip TLS certificate verification
+
+Global Flags:
+      --blob string   File path of device credential blob
+      --debug         Print HTTP contents
+      --tpm string    Use a TPM at path for device credential secrets
+
+$ ./fdo_client onboard -h
+Run FDO TO1 and TO2 onboarding
+
+Usage:
+  fdo_client onboard [flags]
+
+Flags:
+      --cipher string     Name of cipher suite to use for encryption (see usage) (default "A128GCM")
+      --di-key string     Key for device credential [options: ec256, ec384, rsa2048, rsa3072]
+      --download string   A dir to download files into (FSIM disabled if empty)
+      --echo-commands     Echo all commands received to stdout (FSIM disabled if false)
+  -h, --help              help for onboard
+      --insecure-tls      Skip TLS certificate verification
+      --kex string        Name of cipher suite to use for key exchange (see usage)
+      --resale            Perform resale
+      --rv-only           Perform TO1 then stop
+      --upload fsVar      List of dirs and files to upload files from, comma-separated and/or flag provided multiple times (FSIM disabled if empty) (default [])
+      --wget-dir string   A dir to wget files into (FSIM disabled if empty)
+
+Global Flags:
+      --blob string   File path of device credential blob
+      --debug         Print HTTP contents
+      --tpm string    Use a TPM at path for device credential secrets
 
 Key types:
   - RSA2048RESTR
@@ -98,14 +125,14 @@ Remove the credential file if it exists:
 rm cred.bin
 ```
 ### Run the FDO Client with DI URL
-Run the FDO client, specifying the DI URL (on linux systems, root is required to properly gather a device identifier):
+Run the FDO client, specifying the DI URL, key type and credentials blob file (on linux systems, root is required to properly gather a device identifier):
 ```
-./fdo_client -di-device-info=gotest -di http://127.0.0.1:8080 -debug
+./fdo_client device-init --di-device-info=gotest --di http://127.0.0.1:8038 --di-key ec256 --debug --blob cred.bin
 ```
 ### Print FDO Client Configuration or Status
 Print the FDO client configuration or status:
 ```
-./fdo_client -print
+./fdo_client print --blob cred.bin
 ```
 
 ## Execute TO0 from FDO Go Server
@@ -114,12 +141,12 @@ TO0 will be completed in the respective Owner and RV.
 ## Optional: Run the FDO Client in RV-Only Mode
 Run the FDO client in RV-only mode:
 ```
-./fdo_client -rv-only -debug
+./fdo_client onboard --rv-only --di-key ec256 --kex ECDH256 --debug --blob cred.bin
 ```
 ### Run the FDO Client for End-to-End (E2E) Testing
 Run the FDO client for E2E testing:
 ```
-./fdo_client -debug
+./fdo_client --debug
 ```
 
 ## Running the FDO Client with TPM
@@ -136,15 +163,15 @@ Ensure `tpm2_tools` is installed on your system.
    ```
 ### Run the FDO Client with DI URL
 Run the FDO client, specifying the DI URL with the TPM resource manager path specified.
-The suppoerted key type and key exchange must always be explicit through the -di-key and -kex flag.:
+The supported key type and key exchange must always be explicit through the -di-key and -kex flag.:
 ```
-./fdo_client -di http://127.0.0.1:8080 -di-device-info=gotest -di-key ec256 -kex ECDH256 -tpm /dev/tpmrm0 -debug
+./fdo_client device-init --di http://127.0.0.1:8080 --di-device-info=gotest --di-key ec256 --kex ECDH256 --tpm /dev/tpmrm0 --debug
 ```
 >NOTE: fdo_client may require elevated privileges. Please use 'sudo' to execute.
 ### Print FDO Client Configuration or Status
 Print the FDO client configuration or status:
 ```
-./fdo_client -tpm /dev/tpmrm0  -print
+./fdo_client print --tpm /dev/tpmrm0
 ```
 
 ## Execute TO0 from FDO Go Server
@@ -153,11 +180,11 @@ TO0 will be completed in the respective Owner and RV.
 ## Optional: Run the FDO Client in RV-Only Mode
 Run the FDO client in RV-only mode:
 ```
-./fdo_client -rv-only -di-key ec256 -kex ECDH256 -tpm /dev/tpmrm0  -debug
+./fdo_client onboard --rv-only --di-key ec256 --kex ECDH256 --tpm /dev/tpmrm0  --debug
 ```
 ### Run the FDO Client for End-to-End (E2E) Testing
 Run the FDO client for E2E testing:
 ```
-./fdo_client -di-key ec256 -kex ECDH256 -tpm /dev/tpmrm0  -debug
+./fdo_client --di-key ec256 --kex ECDH256 --tpm /dev/tpmrm0  --debug
 ```
