+++ cmd/credential.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: (C) 2024 Intel Corporation
 // SPDX-License-Identifier: Apache 2.0
 
-package main
+package cmd
 
 import (
 	"bytes"
@@ -118,39 +118,39 @@ func readCred() (_ *fdo.DeviceCredential, hmacSha256, hmacSha384 hash.Hash, key 
 		nil,
 		nil
 }
-func loadDeviceStatus(state *FdoDeviceState) bool {
-	if state == nil {
-		return false
-	}
+
+func loadDeviceStatus() (FdoDeviceState, error) {
 	var dataSize int
 	if tpmPath != "" {
 		nv := tpm2.TPMHandle(FDO_CRED_NV_IDX)
 		dataSize = (int)(tpmnv.TpmNVGetSize(tpmc, nv))
+		if dataSize != 0 {
+			var dc fdoTpmDeviceCredential
+			if err := readTpmCred(&dc); err != nil {
+				return FDO_STATE_PC, err
+			}
+			return dc.State, nil
+		}
 	} else {
 		blobData, err := os.ReadFile(filepath.Clean(blobPath))
 		if err != nil {
 			if os.IsNotExist(err) {
 				slog.Debug("DeviceCredential file does not exist. Set state to run DI")
-				*state = FDO_STATE_PRE_DI
-				return true // Return true if the file does not exist
+				return FDO_STATE_PRE_DI, nil
 			}
-			slog.Error("error reading blob credential %q: %v", blobPath, err)
-			return false
+			return FDO_STATE_PC, fmt.Errorf("error reading blob credential %q: %v", blobPath, err)
 		}
-		dataSize = len(blobData)
+		if len(blobData) > 0 {
+			var dc fdoDeviceCredential
+			if err := readCredFile(&dc); err != nil {
+				return FDO_STATE_PC, err
+			}
+			return dc.State, nil
+		}
 	}
 
-	if dataSize == 0 {
-		slog.Debug("DeviceCredential is empty. Set state to run DI")
-		*state = FDO_STATE_PRE_DI
-	} else if resale {
-		*state = FDO_STATE_RESALE
-		slog.Debug("DeviceCredential is non-empty. Set state to resale")
-	} else {
-		slog.Debug("DeviceCredential is non-empty. Set state to run TO1/TO2")
-		// No Device state is being set currently
-	}
-	return true
+	slog.Debug("DeviceCredential is empty. Set state to run DI")
+	return FDO_STATE_PRE_DI, nil
 }
 
 func readCredFile(v any) error {
