+++ go.mod
@@ -3,12 +3,11 @@ module github.com/fido-device-onboard/go-fdo-client
 go 1.23.0
 
 require (
-	github.com/fido-device-onboard/go-fdo v0.0.0-20250512135234-b46a4b0731f2
-	github.com/fido-device-onboard/go-fdo/fsim v0.0.0-20250512135234-b46a4b0731f2
-	github.com/fido-device-onboard/go-fdo/tpm v0.0.0-20250512135234-b46a4b0731f2
+	github.com/fido-device-onboard/go-fdo v0.0.0-20250601145551-2778c77d4779
+	github.com/fido-device-onboard/go-fdo/fsim v0.0.0-20250601145551-2778c77d4779
+	github.com/fido-device-onboard/go-fdo/tpm v0.0.0-20250601145551-2778c77d4779
 	github.com/google/go-tpm v0.9.5
 	github.com/spf13/cobra v1.9.1
-	github.com/spf13/pflag v1.0.6
 	hermannm.dev/devlog v0.5.0
 )
 
@@ -16,6 +15,7 @@ require (
 	github.com/google/go-tpm-tools v0.4.4 // indirect
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/neilotoole/jsoncolor v0.7.1 // indirect
+	github.com/spf13/pflag v1.0.6 // indirect
 	golang.org/x/sys v0.33.0 // indirect
 	golang.org/x/term v0.32.0 // indirect
 )
