+++ main.go
@@ -1,0 +1,16 @@
+// SPDX-FileCopyrightText: (C) 2024 Intel Corporation
+// SPDX-License-Identifier: Apache 2.0
+
+package main
+
+import (
+	"os"
+	"github.com/fido-device-onboard/go-fdo-client/cmd"
+)
+
+func main() {
+	err := cmd.Execute()
+	if err != nil {
+		os.Exit(1)
+	}
+}
